Core: Reorganize test files
They now match the directory structure of the repository
