+++ tests/modules/base/test_admin.py
